chore: split stack trace logic into separate module (#1854)
