+++ crates/blockifier/src/execution.rs
@@ -9,4 +9,5 @@ pub mod entry_point_execution;
 pub mod errors;
 pub mod execution_utils;
 pub mod hint_code;
+pub mod stack_trace;
 pub mod syscalls;
